pyproject.toml
[tool.poetry]
name = "authenticate-users-with-fast-api"
version = "0.1.0"
description = ""
authors = ["Aleks <aleksandar.trpkovski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
